requirements.txt
flask
python-docx
OpenAI
